# openapi.yaml
openapi: 3.0.3
info:
  title: CSCM Tool API
  description: >
    Self-contained REST API for provisioning and managing Minecraft servers as
    Docker containers, with PlayIT tunnel and Cloudflare DNS automation.
  version: "2.0.0"
servers:
  - url: http://localhost:5000
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    ServerId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
paths:
...
  /api/servers/{id}/files:
    get:
      tags: [Files]
      parameters:
        - { $ref: "#/components/parameters/ServerId" }
        - name: path
          in: query
          schema: { type: string }
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
  /api/servers/{id}/files/download:
    get:
      tags: [Files]
      parameters:
        - { $ref: "#/components/parameters/ServerId" }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File contents
        "404": { description: Not found }
  /api/servers/{id}/files/upload:
    post:
      tags: [Files]
      parameters:
        - { $ref: "#/components/parameters/ServerId" }
        - name: path
          in: query
          schema: { type: string }
        - name: filename
          in: query
          schema: { type: string }
          description: Optional target filename inside the destination directory.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        {
          "201": { description: Uploaded },
          "400": { description: Invalid request },
        }
  /api/servers/{id}/files/delete:
    delete:
      tags: [Files]
      parameters:
        - { $ref: "#/components/parameters/ServerId" }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        { "200": { description: OK }, "404": { description: Not found } }
...